project/requirements.txt
numpy==1.26.4
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.2
Faker==25.9.2
Flask==3.0.3
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
MyApplication==0.1.0
packaging==24.1
pandas==2.1.4
pluggy==1.5.0
pycountry==24.6.1
pymongo==4.8.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
tomli==2.0.1
tzdata==2024.2
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
fastapi==0.110.0
uvicorn==0.34.0
psycopg2-binary==2.9.6
SQLAlchemy
joblib==1.3.2
matplotlib==3.6.3
scikit-learn==1.6.1
